Add clip downloader and refiner
Some twitch clips don't appear to have audio in Mac's iMovie (at least I
think that's what it was). Converting the clips to themselves made it
work, I'm not sure how twitch is exporting these files to not have this,
but refining them with refine.sh fixed them. hopefully in premiere pro
this won't be a problem.