Use hashes to detect duplicate clips
Previously if clips had the same title, the second clip (or more) would
not download since it had the same title. Now we download the second (or
more) file and then compare its hash with the previously downloaded one
to detect if it is a duplicate. If it is, we remove the second file
downloaded.