Fix Oauth token issue; organize clips into folders
Oauth started to be required for the endpoint I was using to get the
mp4 information of a clip. Thankfully enough information is returned in
a clip's metadata to correctly determine the mp4 link.

Clips are also now organized into folders based on the week and then
game they were clipped in.